DockerScan/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Docker_Scan/scanner.py: 3
docker == 5.0.0

# Docker_Scan/scanner.py: 12
mysql_connector == 2.2.9

# Docker_Scan/scanner.py: 6
requests == 2.25.1
